#312 Refactor: response에 Id 추가
